+++ src/main/java/com/therapp/spring/controladores/MensajeController.java
@@ -1,17 +1,22 @@
 package com.therapp.spring.controladores;
 
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
 import com.therapp.spring.dto.MensajeDTO;
 import com.therapp.spring.modelo.Mensaje;
 import com.therapp.spring.servicios.MensajeService;
 import com.therapp.spring.servicios.MultimediaMensajeService;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import org.springframework.http.MediaType;
-import org.springframework.web.multipart.MultipartFile;
-import java.util.Collections;
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/messages")
@@ -25,7 +30,7 @@ public class MensajeController {
 
     // GET: listar mensajes entre 2 usuarios
     @GetMapping("/chat/{id1}/{id2}")
-    public ResponseEntity<List<MensajeDTO>> obtenerChat(@PathVariable Integer id1, @PathVariable Integer id2) {
+    public ResponseEntity<List<MensajeDTO>> obtenerChat(@PathVariable Long id1, @PathVariable Long id2) {
         List<MensajeDTO> mensajes = mensajeService.obtenerChat(id1, id2);
         return ResponseEntity.ok(mensajes);
     }
@@ -33,8 +38,8 @@ public class MensajeController {
     // POST: enviar mensaje de id1 -> id2
     @PostMapping(value = "/chat/{id1}/{id2}", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
     public ResponseEntity<MensajeDTO> enviarMensaje(
-        @PathVariable Integer id1,
-        @PathVariable Integer id2,
+        @PathVariable Long id1,
+        @PathVariable Long id2,
         @RequestParam(value = "contenido", required = false) String contenido,
         @RequestParam(value = "archivo", required = false) MultipartFile archivo) {
