+++ src/main/java/com/therapp/spring/dto/MensajeDTO.java
@@ -9,17 +9,17 @@ import lombok.Setter;
 @Getter
 @Setter
 public class MensajeDTO {
-    private int id;
+    private Long id;
     private String contenido;
     private Date fechaEnvio;
     private Boolean visto;
-    private int emisorId;
+    private Long emisorId;
     private String emisorNombre;
-    private int receptorId;
+    private Long receptorId;
     private String receptorNombre;
     private String archivoUrl;
 
-    public MensajeDTO(int id, String contenido, Date fechaEnvio, Boolean visto, int emisorId, String emisorNombre, int receptorId, String receptorNombre, String archivoUrl) {
+    public MensajeDTO(Long id, String contenido, Date fechaEnvio, Boolean visto, Long emisorId, String emisorNombre, Long receptorId, String receptorNombre, String archivoUrl) {
         this.id = id;
         this.contenido = contenido;
         this.fechaEnvio = fechaEnvio;
