+++ src/main/java/com/therapp/spring/repositorios/LocalizacionRepository.java
@@ -3,6 +3,6 @@ package com.therapp.spring.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.therapp.spring.modelo.Localizacion;
 
-public interface LocalizacionRepository extends JpaRepository<Localizacion, Integer> {
+public interface LocalizacionRepository extends JpaRepository<Localizacion, Long> {
     
 }
