+++ src/main/java/com/therapp/spring/repositorios/MultimediaMensajeRepository.java
@@ -3,6 +3,6 @@ package com.therapp.spring.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.therapp.spring.modelo.MultimediaMensaje;
 
-public interface MultimediaMensajeRepository extends JpaRepository<MultimediaMensaje, Integer> {
+public interface MultimediaMensajeRepository extends JpaRepository<MultimediaMensaje, Long> {
 
 }
