+++ src/main/java/com/therapp/spring/repositorios/OrganizacionesRepository.java
@@ -3,6 +3,6 @@ package com.therapp.spring.repositorios;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.therapp.spring.modelo.Organizacion;
 
-public interface OrganizacionesRepository extends JpaRepository<Organizacion, Integer> {
+public interface OrganizacionesRepository extends JpaRepository<Organizacion, Long> {
     
 }
