+++ src/main/java/com/therapp/spring/repositorios/SeguidorRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.therapp.spring.modelo.Seguidor;
 import com.therapp.spring.modelo.Usuario;
 
-public interface SeguidorRepository extends JpaRepository<Seguidor, Integer> {
+public interface SeguidorRepository extends JpaRepository<Seguidor, Long> {
     Optional<Seguidor> findByUsuarioSeguidorAndUsuarioSeguido(Usuario usuarioSeguidor, Usuario usuarioSeguido);
 }
